requirements.txt
asgiref==3.2.10
certifi==2020.6.20
Django==3.0.7
django-appconf==1.0.4
django-crispy-forms==1.9.2
django-imagekit==4.0.2
django-leaflet==0.27.0
django-starcross-gallery==1.0.8
djangorestframework==3.11.0
gunicorn==20.0.4
pilkit==2.0
Pillow==7.2.0
psycopg2==2.7.4
psycopg2-binary==2.8.5
pytz==2020.1
six==1.15.0
sqlparse==0.3.1
whitenoise==5.1.0
wincertstore==0.2
